move enemy in area selection to own file
